+++ src/UI/Components/Diagram/components/ComposerActions.tsx
@@ -9,7 +9,7 @@ import { DependencyContext } from "@/UI/Dependency";
 import { words } from "@/UI/words";
 import { ToastAlert } from "../../ToastAlert";
 import { CanvasContext, InstanceComposerContext } from "../Context/Context";
-import { getServiceOrderItems } from "../helpers";
+import { getServiceOrderItems } from "../helpers/misc";
 import { SavedCoordinates } from "../interfaces";
 import { StyledButton } from "./RightSidebar";
