+++ src/UI/Components/Diagram/halo.ts
@@ -1,5 +1,6 @@
 import { dia, highlighters, ui } from "@inmanta/rappid";
-import { checkIfConnectionIsAllowed, toggleLooseElement } from "./helpers";
+import { checkIfConnectionIsAllowed } from "./helpers/connections";
+import { toggleLooseElement } from "./helpers/visual";
 import { ActionEnum, ConnectionRules, EventActionEnum } from "./interfaces";
 import { ServiceEntityBlock } from "./shapes";
