+++ src/UI/Components/Diagram/paper/paper.ts
@@ -1,11 +1,8 @@
 import { dia, shapes } from "@inmanta/rappid";
 import { anchorNamespace } from "../anchors";
 import createHalo from "../halo";
-import {
-  checkIfConnectionIsAllowed,
-  showLinkTools,
-  toggleLooseElement,
-} from "../helpers";
+import { checkIfConnectionIsAllowed } from "../helpers/connections";
+import { showLinkTools, toggleLooseElement } from "../helpers/visual";
 import collapseButton from "../icons/collapse-icon.svg";
 import expandButton from "../icons/expand-icon.svg";
 import {
